+++ test/next/jdbc/connection_test.clj
@@ -6,7 +6,8 @@
 
   At some point, the datasource/connection tests should probably be extended
   to accept EDN specs from an external source (environment variables?)."
-  (:require [clojure.string :as str]
+  (:require [clojure.java.data :refer [to-java]]
+            [clojure.string :as str]
             [clojure.test :refer [deftest is testing]]
             [next.jdbc.connection :as c]
             [next.jdbc.protocols :as p]))
@@ -103,4 +104,12 @@
           (is (instance? java.sql.Connection con)))))
     (testing "connection via map (Object)"
       (with-open [con (p/get-connection db {})]
+        (is (instance? java.sql.Connection con))))
+    (testing "connection via HikariCP"
+      (with-open [con (p/get-connection (c/jdbc-url com.zaxxer.hikari.HikariDataSource db)
+                                        {})]
+        (is (instance? java.sql.Connection con))))
+    (testing "connection via c3p0"
+      (with-open [con (p/get-connection (c/jdbc-url com.mchange.v2.c3p0.ComboPooledDataSource db)
+                                        {})]
         (is (instance? java.sql.Connection con))))))